--- Annotation/GUI/annotations/django.json
@@ -1,1 +1,0 @@
-{"traffic_light_0": {"x1": 249, "y1": 137, "x2": 389, "y2": 355, "state": 2}, "traffic_light_1": {"x1": 123, "y1": 231, "x2": 231, "y2": 333, "state": 4}}
