+++ Annotation/GUI/annotool/users/Magnus.json
@@ -1,0 +1,5 @@
+{
+    "input_folder": "inputdsfgmh",
+    "output_folder": "..\\test",
+    "output_size": "640x480"
+}
